# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HanapCare Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Doctor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Add a doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorInput"
      responses:
        "201":
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUpdate"
      responses:
        "200":
          description: Updated doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"

  # ─── DEPARTMENTS ─────────────────────────────────────────────
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List consultations
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Create consultation / SOAP note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationInput"
      responses:
        "201":
          description: Consultation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get consultation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
    patch:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /vital-signs:
    get:
      operationId: listVitalSigns
      tags: [consultations]
      summary: List vital signs for a patient
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vital signs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalSign"
    post:
      operationId: createVitalSign
      tags: [consultations]
      summary: Record vital signs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VitalSignInput"
      responses:
        "201":
          description: Vital signs recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalSign"
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [consultations]
      summary: List prescriptions
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [consultations]
      summary: Create a prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # ─── LABORATORY ──────────────────────────────────────────────
  /lab-requests:
    get:
      operationId: listLabRequests
      tags: [lab]
      summary: List lab requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lab requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabRequest"
    post:
      operationId: createLabRequest
      tags: [lab]
      summary: Create lab request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabRequestInput"
      responses:
        "201":
          description: Lab request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
  /lab-requests/{id}:
    get:
      operationId: getLabRequest
      tags: [lab]
      summary: Get lab request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
    patch:
      operationId: updateLabRequest
      tags: [lab]
      summary: Update lab request / add result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabRequestUpdate"
      responses:
        "200":
          description: Updated lab request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"

  # ─── PHARMACY ────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ─── AUDIT LOGS ──────────────────────────────────────────────
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    Doctor:
      type: object
      required: [id, firstName, lastName, licenseNumber, specialization, departmentId]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        specialization:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        isActive:
          type: boolean
    DoctorInput:
      type: object
      required: [firstName, lastName, licenseNumber, specialization, departmentId]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        specialization:
          type: string
        departmentId:
          type: integer
        contactNumber:
          type: string
        email:
          type: string
        availability:
          type: string
    DoctorUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        specialization:
          type: string
        departmentId:
          type: integer
        contactNumber:
          type: string
        email:
          type: string
        availability:
          type: string
        isActive:
          type: boolean

    # ─── DEPARTMENTS ───────────────────────────────────────────
...
    Consultation:
      type: object
      required: [id, patientId, doctorId, consultationDate]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        consultationDate:
          type: string
        subjective:
          type: ["string", "null"]
        objective:
          type: ["string", "null"]
        assessment:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        icdCode:
          type: ["string", "null"]
        status:
          type: string
    ConsultationInput:
      type: object
      required: [patientId, doctorId, consultationDate]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        consultationDate:
          type: string
        subjective:
          type: string
        objective:
          type: string
        assessment:
          type: string
        plan:
          type: string
        diagnosis:
          type: string
        icdCode:
          type: string
    ConsultationUpdate:
      type: object
      properties:
        subjective:
          type: string
        objective:
          type: string
        assessment:
          type: string
        plan:
          type: string
        diagnosis:
          type: string
        icdCode:
          type: string
        status:
          type: string
    VitalSign:
      type: object
      required: [id, patientId, recordedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        recordedBy:
          type: ["string", "null"]
        recordedAt:
          type: string
        bloodPressureSystolic:
          type: ["integer", "null"]
        bloodPressureDiastolic:
          type: ["integer", "null"]
        heartRate:
          type: ["integer", "null"]
        respiratoryRate:
          type: ["integer", "null"]
        oxygenSaturation:
          type: ["number", "null"]
        temperature:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
    VitalSignInput:
      type: object
      required: [patientId, recordedAt]
      properties:
        patientId:
          type: integer
        recordedAt:
          type: string
        bloodPressureSystolic:
          type: integer
        bloodPressureDiastolic:
          type: integer
        heartRate:
          type: integer
        respiratoryRate:
          type: integer
        oxygenSaturation:
          type: number
        temperature:
          type: number
        weight:
          type: number
        height:
          type: number
    Prescription:
      type: object
      required: [id, patientId, doctorId, medicineId, dosage, prescribedAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        medicineId:
          type: integer
        medicineName:
          type: ["string", "null"]
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: ["string", "null"]
        prescribedAt:
          type: string
        status:
          type: string
    PrescriptionInput:
      type: object
      required: [patientId, doctorId, medicineId, dosage, frequency, duration]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        medicineId:
          type: integer
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: string

    # ─── LAB ───────────────────────────────────────────────────
    LabRequest:
      type: object
      required: [id, patientId, doctorId, testType, requestedAt, status]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        testType:
          type: string
        status:
          type: string
        requestedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        resultSummary:
          type: ["string", "null"]
        resultFileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    LabRequestInput:
      type: object
      required: [patientId, doctorId, testType]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        testType:
          type: string
        notes:
          type: string
    LabRequestUpdate:
      type: object
      properties:
        status:
          type: string
        resultSummary:
          type: string
        resultFileUrl:
          type: string
        completedAt:
          type: string

    # ─── PHARMACY ──────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, title, message, type, isRead, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── AUDIT LOGS ────────────────────────────────────────────
    AuditLog:
      type: object
      required: [id, action, userId, userName, timestamp]
      properties:
        id:
          type: integer
        action:
          type: string
        userId:
          type: integer
        userName:
          type: string
        tableName:
          type: ["string", "null"]
        recordId:
          type: ["integer", "null"]
        ipAddress:
          type: ["string", "null"]
        timestamp:
          type: string
        details:
          type: ["string", "null"]