# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HanapCare Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Prescription {
  id: number;
  patientId: number;
  /** @nullable */
  patientName?: string | null;
  doctorId: number;
  /** @nullable */
  doctorName?: string | null;
  medicineId: number;
  /** @nullable */
  medicineName?: string | null;
  dosage: string;
  frequency?: string;
  duration?: string;
  /** @nullable */
  instructions?: string | null;
  prescribedAt: string;
  status?: string;
}